scripts/pi/requirements_audio.txt
# Bender Audio Pipeline - Dépendances Python
# Installation : pip install -r requirements_audio.txt

# Audio core
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0

# VAD WebRTC
webrtcvad>=2.0.10

# MQTT
paho-mqtt>=1.6.0

# Utilitaires
aiofiles>=23.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Logging avancé (optionnel)
structlog>=23.0.0

# Monitoring (optionnel)
psutil>=5.9.0